Life/Death improvements

+++ code/modules/mining/minebot.dm
@@ -92,12 +92,14 @@
 	..()
 
 /mob/living/simple_animal/hostile/mining_drone/death()
-	..()
+	// Only execute the below if we successfully died
+	. = ..()
+	if(!.)
+		return FALSE
 	visible_message("<span class='danger'>[src] is destroyed!</span>")
 	new /obj/effect/decal/cleanable/blood/gibs/robot(src.loc)
 	DropOre(0)
 	qdel(src)
-	return
 
 /mob/living/simple_animal/hostile/mining_drone/attack_hand(mob/living/carbon/human/M)
 	if(M.a_intent == INTENT_HELP)
